♻️ Refactor: Mapper기능에 맞지않는 유틸리티 관련 메서드 관심사 분리
